% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffel_colours.R
\docType{data}
\name{ffel_colours}
\alias{ffel_colours}
\title{Define FFEL Colours}
\format{
An object of class \code{character} of length 4.
}
\usage{
ffel_colours
}
\description{
Any colour adjustments or additions must be done in this vector.
}
\examples{
scales::show_col(ffel_colours)
}
\keyword{datasets}
