% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffel_colours.R
\name{ffel_pal}
\alias{ffel_pal}
\title{Generate FFEL Palettes}
\usage{
ffel_pal(palette = "main", ...)
}
\arguments{
\item{palette}{Name of palette.}

\item{...}{Additional arguments}
}
\description{
Generate palettes of any number of colours using the base palettes in \code{\link{ffel_palettes}}.
}
\details{
Returns a \code{\link{colorRampPalette}} function that creates a vector of colours based on the specified palette. Additional colours are interpolated if the number of colours in the palette is smaller than the number of colours required.
}
