% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffel_colours.R
\docType{data}
\name{ffel_palettes}
\alias{ffel_palettes}
\title{Define FFEL Palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
ffel_palettes
}
\description{
Palettes must be created using the colours in \code{\link{ffel_colours}}.
}
\keyword{datasets}
