% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffel_colours.R
\name{scale_colour_ffel}
\alias{scale_colour_ffel}
\title{Colour Scale Constructor for FFEL Colours}
\usage{
scale_colour_ffel(palette = "main", discrete = TRUE, ...)
}
\arguments{
\item{palette}{Character name of palette from \code{\link{ffel_palettes}}}

\item{discrete}{If TRUE, palette is discrete, if FALSE, palette is continuous}

\item{...}{Additional arguments passed to \code{\link{discrete_scale}} when \emph{discrete = TRUE} or \code{\link{scale_color_gradientn}} when \emph{discrete = FALSE}.}
}
\description{
Apply a custom FFEL colour palette to a \pkg{ggplot2} colour object.
}
\details{
A colour palette based on the FFEL colours. When discrete is FALSE and there are more categories than in the palette, a gradient will be generated.
}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg, colour = as.factor(carb))) +
    geom_point(size = 4) +
    scale_colour_ffel() +
    theme_ffel()

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  geom_point(size = 4) +
  scale_colour_ffel(discrete = FALSE, palette = "two") +
  theme_ffel()
}
