% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffel_colours.R
\name{scale_fill_ffel}
\alias{scale_fill_ffel}
\title{Fill Scale Constructor for FFEL Colours}
\usage{
scale_fill_ffel(palette = "main", discrete = TRUE, ...)
}
\arguments{
\item{palette}{Character name of palette from \code{\link{ffel_palettes}}}

\item{discrete}{If TRUE, palette is discrete, if FALSE, palette is continuous}

\item{...}{Additional arguments passed to \code{\link{discrete_scale}} when \emph{discrete = TRUE} or \code{\link{scale_color_gradientn}} when \emph{discrete = FALSE}.}
}
\description{
Apply a custom FFEL colour palette to a \pkg{ggplot2} fill object.
}
\details{
A colour palette based on the FFEL colours. When discrete is FALSE and there are more categories than in the palette, a gradient will be generated.
}
\examples{
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  scale_fill_ffel(palette = "main") +
  theme_ffel() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

ggplot(mtcars, aes(as.factor(carb), fill = as.factor(carb))) +
  geom_bar() +
  scale_fill_ffel() +
  theme_ffel()
}
