% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ffel.R
\name{theme_ffel}
\alias{theme_ffel}
\title{Custom FFEL \pkg{ggplot2} theme}
\usage{
theme_ffel(base_size = 16)
}
\arguments{
\item{base_size}{Base font size of plot text.}
}
\description{
Applies a custom FFEL theme to a \pkg{ggplot2} object.
}
\details{
This custom theme adjusts a \pkg{ggplot2} object to create report-ready plots. Font sizing is appropriate for a plot approximately 6" square, or similar. For larger or smaller plots, use \emph{base_size} to adjust the font size as necessary to create plots with readable text. All theme adjustments can be over-ridden with subsequent \code{\link{theme}} statements.
}
\examples{
ggplot(mtcars, aes(mpg, disp)) +
  geom_point() +
  theme_ffel()
}
